# Conflicts:
#	presentation/src/main/java/daily/dayo/presentation/screen/main/MainNavigator.kt
#	presentation/src/main/java/daily/dayo/presentation/screen/main/MainScreen.kt
#	presentation/src/main/java/daily/dayo/presentation/screen/settings/SettingsNavigation.kt
#	presentation/src/main/java/daily/dayo/presentation/screen/settings/SettingsScreen.kt
#	presentation/src/main/java/daily/dayo/presentation/view/Button.kt
#	presentation/src/main/res/navigation/nav_graph.xml

+++ data/src/main/java/daily/dayo/data/datasource/remote/heart/HeartResponse.kt
@@ -6,7 +6,7 @@ data class CreateHeartResponse(
     @SerializedName("memberId")
     val memberId: String,
     @SerializedName("postId")
-    val postId: Int,
+    val postId: Long,
     @SerializedName("allCount")
     val allCount: Int
 )
@@ -27,7 +27,7 @@ data class ListAllMyHeartPostResponse(
 
 data class MyHeartPostDto(
     @SerializedName("postId")
-    val postId: Int,
+    val postId: Long,
     @SerializedName("thumbnailImage")
     val thumbnailImage: String
 )
